수정된 도메인에 따른 service 변경사항 반영
