+++ src/main/java/com/cau/vostom/music/service/MusicService.java
@@ -3,17 +3,11 @@ package com.cau.vostom.music.service;
 import com.cau.vostom.music.domain.Music;
 import com.cau.vostom.music.dto.request.DeleteMusicDto;
 import com.cau.vostom.music.dto.request.UploadMusicDto;
-import com.cau.vostom.music.dto.response.ResponseMusicCommentDto;
 import com.cau.vostom.music.repository.MusicRepository;
 import com.cau.vostom.team.domain.Team;
 import com.cau.vostom.team.domain.TeamMusic;
 import com.cau.vostom.team.repository.TeamMusicRepository;
 import com.cau.vostom.team.repository.TeamRepository;
-import com.cau.vostom.team.repository.TeamUserRepository;
-import com.cau.vostom.user.domain.Comment;
-import com.cau.vostom.user.domain.User;
-import com.cau.vostom.user.dto.response.ResponseMyCommentDto;
-import com.cau.vostom.user.repository.CommentRepository;
 import com.cau.vostom.util.api.ResponseCode;
 import com.cau.vostom.util.exception.MusicException;
 import com.cau.vostom.util.exception.UserException;
@@ -21,16 +15,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
 @Service @RequiredArgsConstructor
 public class MusicService {
 
     private final MusicRepository musicRepository;
     private final TeamMusicRepository teamMusicRepository;
     private final TeamRepository teamRepository;
-    private final CommentRepository commentRepository;
 
     //그룹에 음악 업로드
     @Transactional
@@ -47,18 +37,6 @@ public class MusicService {
     public void deleteMusic(DeleteMusicDto deleteMusicDto) {
         Music music = getMusicById(deleteMusicDto.getMusicId());
         musicRepository.delete(music);
-    }
-
-    //노래의 댓글 조회
-    @Transactional(readOnly = true)
-    public List<ResponseMusicCommentDto> getMusicComment(Long musicId) {
-        Music music = getMusicById(musicId);
-        List<Comment> comments = commentRepository.findAllByMusicId(music.getId());
-        if(comments.isEmpty()) { //댓글이 없는 경우
-            return List.of();
-        }
-        return comments.stream().map(ResponseMusicCommentDto::from).collect(Collectors.toList());
-
     }
 
     //음악 학습 요청
